Fix query parameters encoding in rewritten target requests
Fixes request rewriting when query parameters contain a value that
matches a percent-encoded value (e.g. `%2F`).

Previously, such values were not encoded when adding query parameters
to the rewritten request, as the method used internally by
`JerseyUriBuilder.queryParam` to encode values
(UriComponent.contextualEncode) is recognizing percent-encoded values
to avoid double-encoding. This led to changing the original
request, and subsequently the remote denying the request with
`SignatureDoesNotMatch` error.

To fix this, we can encode all query parameter values before passing
them to the URI builder, as we know that they require encoding.